feat: widen LP relaxation presence columns to Long

+++ klause/src/commonTest/kotlin/com/eignex/klause/factor/arithmetic/LinearLinearizerTest.kt
@@ -45,7 +45,7 @@ class LinearLinearizerTest {
         override fun hullEnabled(): Boolean = true
         override fun intColumn(intVar: Int): Int = error("unused")
         override fun boolColumn(boolVar: Int): Int = error("unused")
-        override fun auxColumn(lo: Long, hi: Long, presence: IntArray?): Int = error("unused")
+        override fun auxColumn(lo: Long, hi: Long, presence: LongArray?): Int = error("unused")
         override fun liveDomain(intVar: Int): IntDomain = error("unused")
         override fun declaredDomain(intVar: Int): IntDomain = error("unused")
         override fun row(columns: IntArray, coeffs: LongArray, op: LinearOp, rhs: Long, contribution: Contribution) =
